finish terminal + move around folders
