+++ gui/public/manifest.json
