--- refactor/sagak/src/App.css
@@ -1,73 +1,0 @@
-body {
-  background-color: #261C2E;
-}
-
-.terminal {
-  font-family: 'Roboto Mono';
-  color: white;
-  width: 700px;
-  font-size: 15px;
-}
-
-.term-outer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-
-.cmd {
-  display: flex;
-  margin-top: 3px;
-}
-
-.ascii {
-  display: flex;
-  align-items: flex-end;
-}
-
-#cat {
-  padding-right: 15px;
-}
-
-.help {
-  display: flex;
-  justify-content: space-between;
-}
-
-.newcommand {
-  margin-top: 3px;
-  margin-bottom: 3px;
-}
-
-#link {
-  color: #f4b5f7;
-}
-
-.prompt {
-  background-color: transparent;
-  border: none;
-  color: #f1e6fc;
-  font-family: 'Roboto Mono';
-  font-size: 15px;
-  width: 500px;
-}
-
-.prompt:focus {
-  outline: none;
-}
-
-#prompt {
-  color: #d8b9fa;
-  padding-right: 10px;
-}
-
-.introduction {
-  padding: 2px;
-  margin-bottom: 25px;
-}
-
-#highlight {
-  color: #d8b9fa;
-  text-shadow: #fffce0 1px 0 5px;
-}
