+++ src/Terminal.js
@@ -1,5 +1,5 @@
-import React, { useState } from 'react';
-import { GetEducation, GetCommands, GetAbout, WrongCommand, available_commands } from './utils/commands';
+import React, { useState, useRef, useEffect } from 'react';
+import { GetEducation, GetCommands, GetAbout, WrongCommand, commands, whoami, sudo, Contact } from './utils/commands';
 
 const cat = `
   _                        
@@ -20,35 +20,44 @@ const cat = `
 `
 
 const os = `
-                          ____   _____ 
-                         / __ \\ / ____|
- ___  __ _  __ _  __ _  | |  | | (___  
-/ __|/ _\` |/ _\` |/ _\` | | |  | |\\___ \\ 
-\\__ \\ (_| | (_| | (_| | | |__| |____) |
-|___/\\__,_|\\__, |\\__,_|  \\____/|_____/ 
-            __/ |                      
-           |___/                       
+                              _     
+                             | |    
+  ___  __ _  __ _  __ _   ___| |__  
+ / __|/ _\` |/ _\` |/ _\` | / __| '_ \\ 
+ \\__ \\ (_| | (_| | (_| |_\\__ \\ | | |
+ |___/\\__,_|\\__, |\\__,_(_)___/_| |_|
+             __/ |                  
+            |___/                                        
 `
 
-const prompt = "root@sagak.se $" 
+const prompt = "guest@sagak.se $" 
 
 export default function Terminal() {
   const [history, setHistory] = useState([]);
   const [cmd, setCmd] = useState('');
+  const promptRef = useRef(null)
 
   const handleCommand = () => {
     setHistory([...history, cmd]);
     setCmd('');
   }
 
+  const scrollToBottom = () => {
+    promptRef.current?.scrollIntoView()
+  }
+
+  useEffect(() => {
+    scrollToBottom()
+  }, [history]);
+
   return (
     <div className="terminal"> 
       <div className="ascii">
         <pre id="cat">{cat}</pre>
-        <pre>{os}</pre>
+        <pre id="sh">{os}</pre>
       </div>
       <div className="introduction">
-        <span id="highlight">welcome!</span> so excited you're here. type <span id="highlight">help</span> to display all available commands.
+        ⭐ <span id="highlight">welcome!</span> so excited you're here. type <span id="highlight">help</span> to display all available commands.
       </div>
       <div className="history">
         {history.map((cmd, i) => {
@@ -59,7 +68,10 @@ export default function Terminal() {
               {cmd == "clear" && setHistory([])}
               {cmd == "education" && GetEducation()}
               {cmd == "about" && GetAbout()}
-              {!available_commands.has(cmd) && WrongCommand()}
+              {cmd == "whoami" && whoami()}
+              {cmd == "contact" && Contact()}
+              {cmd.substr(0,4) == "sudo" && sudo()}
+              {!commands.has(cmd) && WrongCommand(cmd)}
             </div>
           )
         })}
@@ -69,6 +81,7 @@ export default function Terminal() {
         <form onSubmit={(e) => {e.preventDefault(); handleCommand();}}>
           <input type="text" value={cmd} onChange={e => setCmd(e.target.value)} className="prompt" autoFocus></input>
         </form>
+        <div ref={promptRef} />
       </div>
     </div>
   );
